chore: Refactor and cleanup
