+++ src/@types/types.ts
@@ -1,0 +1,28 @@
+export enum ProductType {
+  Uncertain = 'uncertain',
+  Wine = 'wine',
+  Beer = 'beer'
+}
+
+export enum RatingResultStatus {
+  NotFound = 'not_found',
+  Uncertain = 'uncertain',
+  Found = 'found'
+}
+
+export type GetRating = {
+  query: ProductType
+  productName: string
+}
+
+export type RatingResponse = {
+  status: RatingResultStatus
+  name: string | null
+  rating: number
+  votes: number
+  link: string | null
+}
+
+export type BeerResponse = RatingResponse & {
+  brewery: string | null
+}
